refactor: Classe `EmpregadoModelo` passa para o pacote models
