+++ src/gui/TelaCadastroEmpregadosController.java
@@ -2,6 +2,7 @@ package gui;
 
 import exceptions.PessoaCPFInvalidoException;
 import exceptions.PessoaDuplicadoException;
+import gui.models.EmpregadoModelo;
 import javafx.beans.property.SimpleFloatProperty;
 import javafx.beans.property.SimpleLongProperty;
 import javafx.beans.property.SimpleStringProperty;
@@ -101,80 +102,5 @@ public class TelaCadastroEmpregadosController {
     public void btnVoltarPressed(ActionEvent event) {
         GerenciadorTelas.getInstance().changeScreen("telaAdmin");
     }
-    //TODO: Possivelmente, tratar como um modelo estático em outra classe separada para GUI
-    public static class EmpregadoModelo {
-        private final SimpleLongProperty uid;
-        private final SimpleStringProperty nome;
-        private final SimpleStringProperty email;
-        private final SimpleFloatProperty reputacao;
-        private final SimpleFloatProperty salarioBase;
 
-        public EmpregadoModelo(Long uid, String nome, String email, Float reputacao, Float salarioBase) {
-            this.uid = new SimpleLongProperty(uid);
-            this.nome = new SimpleStringProperty(nome);
-            this.email = new SimpleStringProperty(email);
-            this.reputacao = new SimpleFloatProperty(reputacao);
-            this.salarioBase = new SimpleFloatProperty(salarioBase);
-        }
-
-        public long getUid() {
-            return uid.get();
-        }
-
-        public SimpleLongProperty uidProperty() {
-            return uid;
-        }
-
-        public void setUid(long uid) {
-            this.uid.set(uid);
-        }
-
-        public String getNome() {
-            return nome.get();
-        }
-
-        public SimpleStringProperty nomeProperty() {
-            return nome;
-        }
-
-        public void setNome(String nome) {
-            this.nome.set(nome);
-        }
-
-        public String getEmail() {
-            return email.get();
-        }
-
-        public SimpleStringProperty emailProperty() {
-            return email;
-        }
-
-        public void setEmail(String email) {
-            this.email.set(email);
-        }
-
-        public float getReputacao() {
-            return reputacao.get();
-        }
-
-        public SimpleFloatProperty reputacaoProperty() {
-            return reputacao;
-        }
-
-        public void setReputacao(float reputacao) {
-            this.reputacao.set(reputacao);
-        }
-
-        public float getSalarioBase() {
-            return salarioBase.get();
-        }
-
-        public SimpleFloatProperty salarioBaseProperty() {
-            return salarioBase;
-        }
-
-        public void setSalarioBase(float salarioBase) {
-            this.salarioBase.set(salarioBase);
-        }
-    }
 }
